Move out file utils to TactilityCore
