+++ Tactility/Include/Tactility/app/filebrowser/FileBrowser.h
@@ -1,14 +1,14 @@
 #pragma once
 
-#include "app/files/View.h"
-#include "app/files/State.h"
+#include "app/filebrowser/View.h"
+#include "app/filebrowser/State.h"
 #include "app/AppManifest.h"
 
 #include <lvgl.h>
 #include <dirent.h>
 #include <memory>
 
-namespace tt::app::files {
+namespace tt::app::filebrowser {
 
 void start();
