+++ Tactility/Source/Tactility.cpp
@@ -37,7 +37,7 @@ namespace app {
     namespace chat { extern const AppManifest manifest; }
     namespace boot { extern const AppManifest manifest; }
     namespace display { extern const AppManifest manifest; }
-    namespace files { extern const AppManifest manifest; }
+    namespace filebrowser { extern const AppManifest manifest; }
     namespace gpio { extern const AppManifest manifest; }
     namespace gpssettings { extern const AppManifest manifest; }
     namespace i2cscanner { extern const AppManifest manifest; }
@@ -78,7 +78,7 @@ static void registerSystemApps() {
     addApp(app::applist::manifest);
     addApp(app::calculator::manifest);
     addApp(app::display::manifest);
-    addApp(app::files::manifest);
+    addApp(app::filebrowser::manifest);
     addApp(app::gpio::manifest);
     addApp(app::gpssettings::manifest);
     addApp(app::i2cscanner::manifest);
